Change SyncMail to SyncAccount use case
