+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/NavigationDrawerModule.kt
@@ -4,7 +4,7 @@ import app.k9mail.feature.navigation.drawer.domain.DomainContract.UseCase
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDisplayAccounts
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDisplayFoldersForAccount
 import app.k9mail.feature.navigation.drawer.domain.usecase.GetDrawerConfig
-import app.k9mail.feature.navigation.drawer.domain.usecase.SyncMail
+import app.k9mail.feature.navigation.drawer.domain.usecase.SyncAccount
 import app.k9mail.feature.navigation.drawer.legacy.AccountsViewModel
 import app.k9mail.feature.navigation.drawer.legacy.FoldersViewModel
 import app.k9mail.feature.navigation.drawer.ui.DrawerViewModel
@@ -37,8 +37,8 @@ val navigationDrawerModule: Module = module {
         )
     }
 
-    single<UseCase.SyncMail> {
-        SyncMail(
+    single<UseCase.SyncAccount> {
+        SyncAccount(
             messagingController = get(),
         )
     }
@@ -66,7 +66,7 @@ val navigationDrawerModule: Module = module {
             getDrawerConfig = get(),
             getDisplayAccounts = get(),
             getDisplayFoldersForAccount = get(),
-            syncMail = get(),
+            syncAccount = get(),
         )
     }
 }
